Fixed Problem 3 (casting to int64).

+++ NumpyIntro/numpy_intro.py
@@ -6,6 +6,7 @@ Thurs. Sept. 15, 2022
 """
 
 import numpy as np
+import os
 
 def prob1():
     """Define the matrices A and B as arrays. Return the matrix product AB."""
@@ -44,7 +45,7 @@ def prob3():
     # 5-6 = -1 on the lower triangle, and 5 above the lower triangle.
     B = np.full((7, 7), 5) - np.tril(np.full((7, 7), 6))
 
-    return A@B@A
+    return np.int64(A@B@A)
 
 
 def prob4(A):
@@ -112,7 +113,7 @@ def prob7():
     adjacent numbers in the same direction (up, down, left, right, or
     diagonally) in the grid. Use slicing, as specified in the manual.
     """
-
+    
     # Load the 20x20 grid.
     grid = np.load('grid.npy')
